ZIGGY-22: upgrade to Gradle 7.5.1.
Gradle doesn't start to support Java 17 until version 7.3, so as a first step we need to move to a more recent Gradle version. The version in this commit works with Gradle 7.5.1, and the Java code should work for either Java 8 or Java 17. The issue with Java is that the jaxb.xml.bind package is removed from the standard Java distribution after Java 8, so all our XML infrastructure needs to move to a different implementation of the JAXB standard. Here we use Jakarta EE 3.0 along with Globefish, which for some reason isn't automatically pulled in by Gradle and Maven dependency management. Anyway, this is definitely a "Han Solo hits the hyperdrive motivator with a wrench" sort of commit, so there's room for improvement.