ZIGGY-22: Code and build that work with Gradle 7.5 and any Java from 8 to 17 (WIP).
Replaced the javax version of XML binding with Jakarta 3.0, which works with Java 8 and Java 17 (and presumably all the versions in between). Fixed some additional Gradle complaints: replacement of JavaExec main with mainClass.set(), fix some implicit dependencies. Eliminate all uses of new Number() (or Number subclasses), as this capability is deprecated in Java 17. At the moment the RMI unit tests fail, so this can't be considered complete as yet.