ZIGGY-22: Fix RMI unit test and exclude RMI intra-process tests.
In another of those "How did this ever work?" situations, RmiInterProcessCommunicationTest class had a failure because it couldn't get the status of the worker (not a surprise, the worker wasn't running). This should always have failed, yet somehow it didn't until now. Fixed by putting in a dummy ClusterController that doesn't try to check the worker status. Also, RmiIntraProcessCommunicationTest and RmiInterProcessCommunicationTest can't both be run in the same test session. We've seen this before on Linux, and it seems to be a test ordering problem: if Inter runs first, Intra bombs, but if Intra runs first, Inter can run. Suspect that gradle test now attempts to execute tests with a deterministic order whereas before it did not. Anyway, all tests run under Java 8 now.