ZIGGY-22: All unit tests now run (though not quite all of them pass) under Java 17 and Gradle 7.5.1.
Fixed a problem in which System.exit() calls were causing the unit tests to terminate prematurely. Updated Mockito to version 3. While I was at it, fixed a couple of deprecated features and moved some classes used to test the RMI system into a common class (they are static inner classes of that class).