ZIGGY-160: test of Up and Next buttons for user manual.
Implemented a couple of navigation features on a single page as a test. One of them is an HTML button; the other is an HTML hyperlink. The hyperlink works when viewing the files in Typora, but the button looks better. We'll have to see how this looks on GitHub, once we've got our package there.

(cherry picked from commit 488fc48e1c2316d47fad6310cc2a80f42e9a6290)

# Conflicts:
#	README.md