ZIGGY-160: test of Up and Next buttons for user manual.
Implemented a couple of navigation features on a single page as a test. One of them is an HTML button; the other is an HTML hyperlink. The hyperlink works when viewing the files in Typora, but the button looks better. We'll have to see how this looks on GitHub, once we've got our package there.