ZIGGY-22: complete Gradle 7 migration.
Codebase now builds under Gradle 7.6 with Java 8, and all unit tests pass. Most of the changes are cleanup from either the jaxb / Jakarta migration or from the merge of dev into this branch. A couple of other changes were needed for unit tests. The old hdf5.gradle plugin is now copy-hdf5.gradle. Rationale: for some reason in this version of Gradle, the build fails when hdf5.gradle comes after database-schemas.gradle in the build.gradle file; this is because when database-schemas is applied, the HDF5 copy tasks it depends on aren't yet defined. The solutions to this are to either move hdf5.gradle up in the list (thus breaking the alpha ordering), or else find a new name for hdf5.gradle that preserves alpha ordering but puts it above database-schemas.gradle. Ta-da!