ZIGGY-180: implement ZiggyUnitTest abstract class and use AlgorithmMonitorTest as example of same.
The ZiggyUnitTest is a new abstract class that supports a more automated approach to setup and teardown of test classes. To demonstrate its benefits, the AlgoithmMonitorTest class is converted to a subclass of ZiggyUnitTest. The AlgorithmMonitorTest is modified to remove all Thread.sleep() calls that had been used to allow the test thread to wait for the ScheduledThreadPoolExecutor to run one or more times. Instead, the AlgorithmMonitor run() method is called directly.