ZIGGY-180: eliminate Thread.sleep() states in ProcessHeartbeatManagerTest.
Removing the sleep states from ProcessHeartbeatManagerTest proved to be a bit more complicated than some of the other cases. For one thing, it became necessary to have a static method that can either provide the actual system time (for normal operations), or a user-supplied value (for testing). Also, the checkForHeartbeat() method had to be refactored so that in normal operations it always tries to reinitialize the heartbeat manager, but in test the reinitialization can be performed separately (to simulate the passage of time while waiting for a new heartbeat to be detected in the initializer).