ZIGGY-180: remove Thread.sleep() from PipelineOutputsTest.
The only reason for the sleep() in PipelineOutputsTest was to wait for file write locks to clear before deleting the build/test directory. At a later point in working this ticket we're going to stop deleting either build/test or its contents, so I simply deleted the sleep() and the directory removal.