ZIGGY-180: first prototype of ZiggyDirectoryRule.
Created a new implementation of JUnit's TestRule for managing directories during unit tests. The rule creates a per-test-class subdir to build/test, and then per-test-method subdirs in the per-test-class subdir. Per-test-method subdirs are cleaned prior to use but not after use, so the volume of temporary files doesn't grow monotonically as the user runs unit tests over and over, but the temporary files produced during execution are preserved for diagnostic use, if needed. Also updated the DefaultDataImporterTest class to use the new Rule.