ZIGGY-120: Delete Thread.sleep() from DataFileManagerTest.
While renovating the DataFileManagerTest class, I discovered that the ZiggyDirectoryRule needs to be able to force the permissions on a directory tree for deletion. Thus the cleanDirectoryTree() and deleteDirectoryTree() methods in FileUtil now take optional arguments for whether or not to force deletion (i.e., force == change permissions to allow deletion).