ZIGGY-180: Update RMI unit tests to eliminate Thread.sleep() calls, plus other improvements.
Eliminated use of Thread.sleep() in RMI unit tests. In doing so, I realized that we have a general need for code that waits for some kind of event to happen or for a timeout to elapse. To this end, I added the TestEventDetector class and used it in the RMI unit test classes. Finally, made Gradle clean depend on a task that sets permissions in build/test to allow build/test to be deleted.