ZIGGY-180: Last bits of cleanup in unit test area.
Removed the MemoryMonitor class that's not used anywhere, along with its unit test class. Adjusted the time limit in the testWaiting() method of PipelineInstanceManagerTest to allow it to pass more reliably. At the moment, the PipelineInstanceManager uses a busy loop to wait between launches of pipeline instances. This should be rewritten at some point.