ZIGGY-197: improvements to recursive permission setters.
Modified the cleanDirectory() method so it can change the permissions of the root directory. Removed the recursive permission setter that takes one permission string argument. Added 2 new methods: one that prepares a directory tree for overwrites; one that write-protects a directory tree. Updated unit tests to include the new methods.