ZIGGY-199: prototype of compute node with no socket use for server / client communication.
Eliminated the use of a socket for communication between the SubtaskServer and the SubtaskClients (which are, in any event, in the same JVM so they really don't need to use sockets). In place of sockets, the SubtaskServer and SubtaskClient use ArrayBlockingQueue instances for their messages; this allows those classes to block when waiting for a new message or new response. In the process, I did a bit of additional cleanup on the exception handling in the ComputeNodeMaster system: removed unneeded throws statements (i.e., for runtime exceptions and also for SocketException), add a method that allows the ComputeNodeMaster to see whether the SubtaskServer's listener is running, remove iterative attempts to send a request from the client to the server (these were needed when a request could get messed up by a transient problem with the socket, but without a socket there's no longer any way for this to happen, hence no need for retries).