Remove version of wrapper in build.gradle
Turns out that we should not use a version for the Wrapper jarfile in build.gradle. This is because we download and install a specific version in wrapper.gradle, so there's only one version available in any event. Also, Gradle gets confused if we tell it to look for a version like "3.5.+".