ZIGGY-199: complete unit tests for socket-free ComputeNodeMaster infrastructure
Added unit tests for the ComputeNodeMaster and SubtaskMaster classes. To do this, it was necessary to add some functions that could be modified by Mockito spy instances. Also, the LockManager method that releases all locks had to be made public so that SubtaskMaster can run it. Otherwise, the locks acquired by SubtaskMasterTest instances can somehow cause the LockManagerTest methods to fail.