ZIGGY-204: restore keep-up processing functionality
Reorganize the methods in the DataFileManager so that they correctly use the TaskConfigurationParameters in the PipelineTask to decide whether to exercise keep-up or full processing. Remove TaskConfigurationParameters arguments from all DataFileManager method arguments (the TaskConfigurationParameters should not vary from one method call to another; the DataFileManager should be able to rely on the copy in its PipelineTask instance for all method calls). Update the DataFileManager methods so that they always make correct decisions about how to filter files based on the keep-up versus full processing configuration parameter.