ZIGGY-207: add standard ZIggy charset (UTF-8)
The most common SpotBugs complaint is an internationalization issue, in which we create things like PrintStreams without explicitly specifying the charset to use. Here we define ZIGGY_CHARSET to be UTF-8, and ZIGGY_CHARSET_NAME to be ZIGGY_CHARSET.name(), so we can address all the instances of this complaint in a uniform way, and one that we can then change later if it turns out that UTF-8 is actually bad.