ZIGGY-207: Fix security audit issues found by SpotBugs
The SpotBugs security plugin identified a handful of issues in Ziggy. For the most part we've had to use a SuppressFBWarnings annotation that documents why we think the issue in question is not a security issue for Ziggy. In a few cases we were able to address the issues by deleting classes or methods that are no longer needed.