ZIGGY-142: WIP, most compilation warnings fixed
This fixes most of the compilation warnings in both src and buildSrc. What's left are 2 issues. First, Hibernate's NamingStrategy is deprecated, and the rules for replacing it are unclear to me at the moment. Second, The AboutDialog uses a deprecated Jgoodies class, and the nominal replacement is totally different, hence I don't know how to get from here to there.