ZIGGY-193: reorganize, and add a few articles to the user manual
Based on discussions with BW, the last 30% or so of the user manual has been reorganized. The catchall advanced topics page is gone. Instead we have some additional top-level articles, plus a new catchall (Ziggy's Dustier Corners, inspired by the chapter of K & R that has a similar title). A few of the new top-level articles are added, along with a couple of lower-level ones.