ZIGGY-234: use the properties file to give runjava the location of the Java executable
Added a property to the pipeline properties, java.home, which performs the same function for Ziggy as the environment variable JAVA_HOME does for the shell. If set, this property is used by runjava to build its command line that calls java. This ensures that the user can specify a particular version Java, located at a particular location, when running runjava. In particular, this is necessary when the ComputeNodeMaster is running, since the compute node that runs ComputeNodeMaster may or may not have the correct JAVA_HOME env var.